feature: remodularization of code, new fields added to match rfc compatibility. Fixes issue #15
