+++ vCardLib/Helpers/Enums.cs
@@ -32,4 +32,14 @@
         Overwrite,
         ThrowError
     }
+
+    /// <summary>
+    /// vCard version
+    /// </summary>
+    public enum Version
+    {
+        V2,
+        V3,
+        V4
+    }
 }
