+++ vCardLib/Serializers/Serializer.cs
@@ -1,0 +1,7 @@
+﻿namespace vCardLib.Serializers
+{
+    public class Serializer
+    {
+
+    }
+}
